refactor : 불필요한 `Manual`, 'Auto' 객체 제거하여 곧바로 생성가능하게 변경
