--- src/main/java/lotto/domain/model/LottoGenerator.java
@@ -1,8 +1,0 @@
-package lotto.domain.model;
-
-public interface LottoGenerator {
-    
-    LottoTickets generateTickets();
-    
-    BuyCount getBuyCount();
-}
