+++ src/main/java/lotto/domain/model/LottoGenertor/LottoGenerator.java
@@ -1,0 +1,11 @@
+package lotto.domain.model.LottoGenertor;
+
+import lotto.domain.model.BuyCount;
+import lotto.domain.model.LottoTickets;
+
+public interface LottoGenerator {
+    
+    LottoTickets generateTickets();
+    
+    BuyCount getBuyCount();
+}
