--- src/main/java/lotto/domain/model/Manual.java
@@ -1,7 +1,0 @@
-package lotto.domain.model;
-
-import java.util.List;
-
-public record Manual(List<Lotto> manualLottoList) {
-    
-}
