+++ src/test/java/lotto/domain/model/LottoGenertor/LottoAutoGeneratorTest.java
@@ -1,8 +1,7 @@
-package lotto.domain.model.impl;
+package lotto.domain.model.LottoGenertor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.domain.model.LottoGenerator;
 import lotto.domain.model.LottoTickets;
 import org.junit.jupiter.api.Test;
