--- src/test/java/lotto/domain/model/ManualTest.java
@@ -1,5 +1,0 @@
-package lotto.domain.model;
-
-class ManualTest {
-    
-}
